+++ dvc/repo/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 
 import os
+
 import dvc.prompt as prompt
 import dvc.logger as logger
