+++ dvc/repo/imp.py
@@ -1,4 +1,4 @@
-from dvc.scm import scm_context
+from dvc.repo.scm_context import scm_context
 
 
 @scm_context
