+++ dvc/repo/run.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-from dvc.scm import scm_context
+from dvc.repo.scm_context import scm_context
 
 
 @scm_context
